# storage/postgreskv: use batch size 1000