@thepaul 2019-05-17
Dang it y'all went too fast. I'm going to revert. I'm getting _very_ consistently the opposite results in my own testing, which is surprising because I can't see any way that defaultBatchSize should even be able to affect those tests at all. So something is off here.