@egonelbre 2019-05-17
Ah, sorry this shouldn't have been "Reviewer Can Merge", we still need benchmarks. +1 for reverting.